requirements.txt
fastapi
uvicorn
sentence-transformers
torch==2.2.0+cpu
faiss-cpu
numpy<2.0
pydantic-settings
psycopg2-binary
